Unity android inject config (#9)
* added ignore to Logs folder

* more gitignores

* added callback methods to androidProxy

* formatted unityProxy

* add setUserId to androidProxy

* added setUserEmail

* buttons for messaging

* ios: in-app API

* fancier log

* Logs and git ignore

* added buttonController

* deleted exampleAppScript

* added eventTracking functions

* added push functionality

* handlers + setting custom modulemap for unity framework

* ios project setup script

* scrollable panel + clarity with positioning

* return InAppButtonPress

* fixed pushed messaging functionality

* enabled in-app messaging

* updated pushHandling

* refactored onPushOpen handling

* added handler setting functions

* added in-app hadlers and summary functionality

* Delete ExampleApp/firstTest_BurstDebugInformation_DoNotShip/tempburstlibs/armeabi-v7a directory

* fixed merge errors

* renamed deepLink in-app handler

* added overrideinstallInfo

* added in-app strategy input check

* renamed eventParams

* removed getCurrentUserId

* reduced nesting

* auto formatting

* added android setting-up

* delete inject config changes

* corrected override data

* removed error

* removed try and catch from setUserId

* configured git ignore

* added files to git ignore

* added optimoveMainActivity

* deleted redundant try and catch

* capitalize handler setters + add setter for ddl handler

* fix compilation error + example of using named delegate

* revert scene changes

* turn off burst compilation

* set in-app deep link handler only when inApp is enabled + correct return

* added intent-filter to manifest

* simplify

* set minSdk to 21

* depend on latest android sdk + AndroidPlugin as a standalone push-integrated project

* consistent indentation

* remove unused method

* consistent formatting + rename aar output

* manifest merged with Firebase package, clean gitignore items

* revert of an incorrect commit (reassign elements to the ButtonController script)

* fixed push register

* fixed pushRegister

* added firebase ios to git ignore

* remove outdated usage examples

* increase priority of our FirebaseMessagingService intent filter

* extended MessagingUnityPlayerActivity

* fix validation errors

* added deep link handler

* git cache cleared

* Delete lib_burst_generated.txt

* deleted unnecessary files

* fixed unwanted merge changes

* update firebase messaging version

* merged android messaging branch

* added pollPendingDdl

* added onNewIntent override

* changed pollPendingPush calling order

* removed firebaseMessaging.unityPackage

* remove example usage

* auto-setup android project without firebase.unitypackage

* replaced MessagingUnityPlayerActivity with UnityPlayerActivity

* changed deepLink subdomain

* removed unityActivity from gradle settings

* deleted firebase messaging unity aar

* delete unnecessary logs

* added conifg injection to res folder

* added optimove.xml

* added getStringConifgValue

* optimove config injection

+++ ExampleApp/Assets/OptimoveSdk/Editor/SetUpGradleProject.cs
@@ -2,11 +2,19 @@ using UnityEditor;
 using UnityEditor.Android;
 using UnityEngine;
 using System.IO;
+using UnityEditor.Build;
 
 class SetUpGradleProject : IPostGenerateGradleAndroidProject
 {
     public int callbackOrder { get { return 0; } }
     public void OnPostGenerateGradleAndroidProject(string pathToUnityLibrary)
+    {
+        MoveGoogleServices(pathToUnityLibrary);
+
+        MoveOptimoveConfig(pathToUnityLibrary);
+    }
+
+    private void MoveGoogleServices(string pathToUnityLibrary)
     {
         string src = Path.Combine(pathToUnityLibrary, "src/main/assets/google-services.json");
         string dest = Path.Combine(pathToUnityLibrary, "../launcher/google-services.json");
@@ -15,6 +23,22 @@ class SetUpGradleProject : IPostGenerateGradleAndroidProject
         }
 
         if (File.Exists(dest)){
+            FileUtil.DeleteFileOrDirectory(dest);
+        }
+
+        FileUtil.MoveFileOrDirectory(src, dest);
+    }
+
+    private void MoveOptimoveConfig(string pathToUnityLibrary) {
+        string src = Path.Combine(pathToUnityLibrary, "src/main/assets/optimove.xml");
+        string dest = Path.Combine(pathToUnityLibrary, "../launcher/src/main/res/values/optimove.xml");
+        if (!File.Exists(src))
+        {
+            throw new BuildFailedException("optimove.xml is missing in StreamingAssets");
+        }
+
+        if (File.Exists(dest))
+        {
             FileUtil.DeleteFileOrDirectory(dest);
         }
